// src/store/ducks/header/types.ts
export enum HEADER_ACTION_TYPES {
  FETCH = 'FETCH',
  FAMILY_SELECTED = 'FAMILY_SELECTED',
  CATEGORY_SELECTED = 'CATEGORY_SELECTED',
  CATEGORY_DETAILS_FETCHED = 'CATEGORY_DETAILS_FETCHED',
  SHOW_SEARCH = 'SHOW_SEARCH',
  HIDE_SEARCH = 'HIDE_SEARCH',
  SHOW_PROFILE_LOGIN = 'SHOW_PROFILE_LOGIN',
  HIDE_PROFILE_LOGIN = 'HIDE_PROFILE_LOGIN',
}