// src/store/ducks/newsletter/types.ts
export enum NEWSLETTER_ACTION_TYPES {
  RESET = 'RESET',
  SUBMIT = 'SUBMIT',
  SUBMITTING = 'SUBMITTING',
  ADDED_SUCCESS = 'ADDED_SUCCESS',
  ADDED_FAILURE = 'ADDED_FAILURE',
}

export enum NEWSLETTER_LINKID_TYPES {
  TRY_FETCH = 'TRY_FETCH',
  FETCHING = 'FETCHING',
  FETCH_SUCCESS = 'FETCH_SUCCESS',
  FETCH_FAILURE = 'FETCH_FAILURE',
  TRY_SUBMIT = 'TRY_SUBMIT',
  SUBMITTING = 'SUBMITTING',
  SUBMIT_SUCCESS = 'SUBMIT_SUCCESS',
  SUBMIT_FAILURE = 'SUBMIT_FAILURE',
}

export enum RECIPE_RATINGS_EMAIL_TYPES {
  TRY_SUBMIT = 'TRY_SUBMIT',
}