// src/store/ducks/recipesBox/types.ts
export enum RECIPEBOX {
  TRY_FETCH_RECIPEBOX = 'TRY_FETCH_RECIPEBOX',
  FETCHING_RECIPEBOX = 'FETCHING_RECIPEBOX',
  FETCHED_RECIPEBOX_SUCCESS = 'FETCHED_RECIPEBOX_SUCCESS',
  FETCHED_RECIPEBOX_FAILURE = 'FETCH_RECIPEBOX_FAILURE',

  TRY_ADD_RECIPE = 'TRY_ADD_RECIPE',
  ADDING_RECIPE = 'ADDING_RECIPE',
  ADD_RECIPE_SUCCESS = 'ADD_RECIPE_SUCCESS',
  ADD_RECIPE_FAILURE = 'ADD_RECIPE_FAILRE',
  TRY_DELETE_RECIPE = 'TRY_DELETE_RECIPE',
  DELETING_RECIPE = 'DELETING_RECIPE',
  DELETE_RECIPE_SUCCESS = 'DELETE_RECIPE_SUCCESS',
  DELETE_RECIPE_FAILURE = 'DETEL_RECIPE_FAILURE',

  TRY_ADD_ARTICLE = 'TRY_ADD_ARTICLE',
  ADDING_ARTICLE = 'ADDING_ARTICLE',
  ADD_ARTICLE_SUCCESS = 'ADD_ARTICLE_SUCCESS',
  ADD_ARTICLE_FAILURE = 'ADD_ARTICLE_FAILURE',
  TRY_DELETE_ARTICLE = 'TRY_DELETE_ARTICLE',
  DELETING_ARTICLE = 'DELETING_ARTICLE',
  DELETE_ARTICLE_SUCCESS = 'DELETE_ARTICLE_SUCCESS',
  DELETE_ARTICLE_FAILURE = 'DELETE_ARTICLE_FAILURE'
}